expression: refactor expression
Firstly, abstract SubQuery to interface{}, some mock test are not able to keep.
